Decouple interactive planning manual test (#585)
