+++ tests/manual/CMakeLists.txt
@@ -35,10 +35,16 @@ target_link_libraries(${test_prefix}kvstore_console mg-kvstore gflags mg-utils)
 add_manual_test(query_hash.cpp)
 target_link_libraries(${test_prefix}query_hash mg-query)
 
-add_manual_test(query_planner.cpp interactive_planning.cpp)
+add_manual_test(query_planner.cpp interactive/planning.cpp)
 target_link_libraries(${test_prefix}query_planner mg-query)
 if (READLINE_FOUND)
   target_link_libraries(${test_prefix}query_planner readline)
+endif()
+
+add_manual_test(query_execution_dummy.cpp)
+target_link_libraries(${test_prefix}query_execution_dummy mg-query)
+if (READLINE_FOUND)
+  target_link_libraries(${test_prefix}query_execution_dummy readline)
 endif()
 
 add_manual_test(expression_pretty_printer.cpp)
