+++ tests/manual/interactive/planning.hpp
@@ -21,20 +21,11 @@
 #include "query/frontend/semantic/symbol_table.hpp"
 #include "query/plan/operator.hpp"
 
+#include "plan.hpp"
+
 namespace database {
 class GraphDbAccessor;
 }
-
-struct InteractivePlan {
-  // Original plan after going only through the RuleBasedPlanner.
-  std::unique_ptr<memgraph::query::plan::LogicalOperator> unoptimized_plan;
-  // Storage for the AST used in unoptimized_plan
-  memgraph::query::AstStorage ast_storage;
-  // Final plan after being rewritten and optimized.
-  std::unique_ptr<memgraph::query::plan::LogicalOperator> final_plan;
-  // Cost of the final plan.
-  double cost;
-};
 
 typedef std::vector<InteractivePlan> PlansWithCost;
