+++ tests/manual/query_planner.cpp
@@ -9,7 +9,7 @@
 // by the Apache License, Version 2.0, included in the file
 // licenses/APL.txt.
 
-#include "interactive_planning.hpp"
+#include "interactive/planning.hpp"
 
 #include <gflags/gflags.h>
